refactor(mcp): split feishu-context-mcp.ts into modular components (Issue #693)
Split the 1259-line feishu-context-mcp.ts into smaller, more maintainable modules:

- src/mcp/feishu-context-mcp.ts: Entry point (~360 lines, reduced from 1259)
- src/mcp/tools/
  - types.ts: Shared type definitions
  - send-message.ts: send_user_feedback tool
  - send-file.ts: send_file_to_feishu tool
  - card-interaction.ts: update_card, wait_for_interaction
  - index.ts: Module exports

- src/mcp/utils/
  - card-validator.ts: Feishu card validation utilities
  - feishu-api.ts: Low-level Feishu API helpers
  - index.ts: Module exports

Benefits:
- Better separation of concerns
- Easier testing of individual components
- Clearer responsibilities
- Improved maintainability
- Backward compatibility maintained

Test Results:
- All 63 MCP tests pass
- TypeScript compilation passes

Fixes #693
