+++ src/mcp/tools/index.ts
@@ -1,0 +1,18 @@
+/**
+ * Tool implementations for Feishu MCP.
+ *
+ * @module mcp/tools
+ */
+
+export type {
+  SendFeedbackResult,
+  SendFileResult,
+  UpdateCardResult,
+  WaitForInteractionResult,
+  MessageSentCallback,
+  PendingInteraction,
+} from './types.js';
+
+export { send_user_feedback, setMessageSentCallback, getMessageSentCallback } from './send-message.js';
+export { send_file_to_feishu } from './send-file.js';
+export { update_card, wait_for_interaction, resolvePendingInteraction } from './card-interaction.js';
