+++ src/mcp/utils/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Utility functions for Feishu MCP.
+ *
+ * @module mcp/utils
+ */
+
+export { isValidFeishuCard, getCardValidationError } from './card-validator.js';
+export { sendMessageToFeishu } from './feishu-api.js';
